+++ TIG/TIG/Data/DefaultWeeklyRepeatRepository.swift
@@ -1,0 +1,28 @@
+//
+//  DefaultWeeklyRepeatRepository.swift
+//  TIG
+//
+//  Created by 이정동 on 7/20/24.
+//
+
+import Foundation
+
+final class DefaultWeeklyRepeatRepository: SwiftDataStorage {
+    
+}
+
+extension DefaultWeeklyRepeatRepository: WeeklyRepeatRepository {
+    func createWeeklyRepeat() {
+        
+    }
+    
+    func fetchWeeklyRepeats() -> [WeeklyRepeat] {
+        return []
+    }
+    
+    func updateWeeklyRepeat() {
+        
+    }
+    
+    
+}
