+++ TIG/TIG/Domain/Interface/DailyContentRepository.swift
@@ -7,15 +7,9 @@
 
 import Foundation
 
-protocol PersistenceRepository {
-    // DailyContent 관련
+protocol DailyContentRepository {
     func createDailyContent()
     func fetchDailyContents() -> [DailyContent]
     func updateDailyContent()
     func deleteDailyContent()
-    
-    // WeeklyRepeat 관련
-    func createWeeklyRepeat()
-    func fetchWeeklyRepeats() -> [WeeklyRepeat]
-    func updateWeeklyRepeat()
 }
