+++ TIG/TIG/Domain/Interface/SettingRepository.swift
@@ -1,0 +1,8 @@
+//
+//  SettingRepository.swift
+//  TIG
+//
+//  Created by 이정동 on 7/20/24.
+//
+
+import Foundation
