+++ TIG/TIG/Domain/Interface/WeeklyRepeatRepository.swift
@@ -1,0 +1,14 @@
+//
+//  WeeklyRepeatRepository.swift
+//  TIG
+//
+//  Created by 이정동 on 7/20/24.
+//
+
+import Foundation
+
+protocol WeeklyRepeatRepository {
+    func createWeeklyRepeat()
+    func fetchWeeklyRepeats() -> [WeeklyRepeat]
+    func updateWeeklyRepeat()
+}
